Making use of InputStream and OutputStream for saving bank information
BankProgram has been modified so that it reads the bank's initial
account information from a file when the program begins,
and writes the account information to the file when the program ends.
It uses the class SavedBankInfo to interact with the file.

The public method nextAcctNum is added to Bank,
so that the value of its nextAcct variable can be saved.