Refactoring to take advantage of The Template Pattern
Due to The Template Pattern, there is no duplicated code whatsoever!

The AbstractBankAccount class has been redesigned to use the
template pattern. In particular, the class now partially implements
the template methods hasEnoughCollateral, toString, and addInterest,
using the helper abstract methods collateralRatio, accountType,
and interestRate.

SavingsAccount is rewritten to implement those abstract methods.

CheckingAccount is rewritten so implement collateralRatio,
and declares accountType and interestRate to be abstract.

RegularChecking and InterestChecking are rewritten to implement
accountType and interestRate.