Adding static factory methods to create accounts
The new interface AccountFactory acts as mediator between
BankAccount's types and Bank.

The new interface AccountFactory contains a static factory method
for each of the three types of bank account:
createSavings, createRegularChecking, and createInterestChecking.
It also has a static method createAccount, which uses the type
number to determine which account to create.

The method createSavingsWithDeposit in BankAccount has been
rewritten to mention AccountFactory.createSavings.

The methods newAccount in Bank and readAccount in SavedBankInfo
are modified to call the createAccount method in AccountFactory.