Taking advantage of The Observer Pattern to make Bank observable
The program is modified to allow Bank to be an observable.
The observer interface is BankObserver.

The bank supports four different event types, which are
defined in the enum BankEvent.

The Bank class is modified to implement a map of BankObserver
observer lists, one for each event type.

Bank has the methods addObserver, removeObserver and notifyObservers.
The Bank methods that update accounts are modified to call notifyObservers.

There is a new class Auditor that implements BankObserver.

BankProgram is modified to add two observers to the bank.